// src/models/appointment.models.ts
import { Payment } from './payment.models';
import { NewAddress } from './newadd.models';
import { Service } from './service.models';
import { User } from "./user.models";
import { Profile } from "./profile.models";
import { Address } from "./address.models";
import { Category } from "./category.models";
import { StatusLog } from "./status-log.models";
import { Subs } from './subs.models';

export class Appointment {
    id: number;
    provider_id: number;
    provider_status: string;
    total: string;
    service: Service;
    subscription: Subs;
    user_id: number;
    address_id: number;
    status: string;
    date: string;
    time_from: string;
    time_to: string;
    date_formatted: string;
    time_from_formatted: string;
    time_to_formatted: string;
    created_at: string;
    updated_at: string;
    user: User;
    provider: Profile;
    address: NewAddress;
    category: Category;
    logs: Array<StatusLog>;
    payment_id: number;
    payment: Payment;
    type: string;
    service_name: string;
    service_id: number;
    pa_id: number;
    pa_status: string;
    c_at: string;
    units: number;
    image_url: string;
    uname: string;
    area: string;
    landmark: string;
    task: string;
    code: number;
    e_payment: string;
    e_amount: string;
    e_status: string;
    cancelled_by: string;
    cancelled_reason: string;
    extend_date: string;
    extend_reason: string;
    time_slot: number;
}